moving static code from NetworkBehaviour to its own class
