+++ Assets/Mirror/Runtime/NetworkBehaviour.cs
@@ -1,6 +1,7 @@
 using System;
 using System.Collections.Generic;
 using System.ComponentModel;
+using Mirror.RemoteCalls;
 using UnityEngine;
 
 namespace Mirror
@@ -167,18 +168,6 @@ namespace Mirror
         }
 
         #region Commands
-
-        internal static int GetMethodHash(Type invokeClass, string methodName)
-        {
-            // (invokeClass + ":" + cmdName).GetStableHashCode() would cause allocations.
-            // so hash1 + hash2 is better.
-            unchecked
-            {
-                int hash = invokeClass.FullName.GetStableHashCode();
-                return hash * 503 + methodName.GetStableHashCode();
-            }
-        }
-
         [EditorBrowsable(EditorBrowsableState.Never)]
         protected void SendCommandInternal(Type invokeClass, string cmdName, NetworkWriter writer, int channelId, bool ignoreAuthority = false)
         {
@@ -209,7 +198,7 @@ namespace Mirror
                 netId = netId,
                 componentIndex = ComponentIndex,
                 // type+func so Inventory.RpcUse != Equipment.RpcUse
-                functionHash = GetMethodHash(invokeClass, cmdName),
+                functionHash = RemoteCallHelper.GetMethodHash(invokeClass, cmdName),
                 // segment to avoid reader allocations
                 payload = writer.ToArraySegment()
             };
@@ -226,7 +215,7 @@ namespace Mirror
         [EditorBrowsable(EditorBrowsableState.Never)]
         public virtual bool InvokeCommand(int cmdHash, NetworkReader reader)
         {
-            return InvokeHandlerDelegate(cmdHash, MirrorInvokeType.Command, reader);
+            return RemoteCallHelper.InvokeHandlerDelegate(cmdHash, MirrorInvokeType.Command, reader, this);
         }
         #endregion
 
@@ -253,7 +242,7 @@ namespace Mirror
                 netId = netId,
                 componentIndex = ComponentIndex,
                 // type+func so Inventory.RpcUse != Equipment.RpcUse
-                functionHash = GetMethodHash(invokeClass, rpcName),
+                functionHash = RemoteCallHelper.GetMethodHash(invokeClass, rpcName),
                 // segment to avoid reader allocations
                 payload = writer.ToArraySegment()
             };
@@ -294,7 +283,7 @@ namespace Mirror
                 netId = netId,
                 componentIndex = ComponentIndex,
                 // type+func so Inventory.RpcUse != Equipment.RpcUse
-                functionHash = GetMethodHash(invokeClass, rpcName),
+                functionHash = RemoteCallHelper.GetMethodHash(invokeClass, rpcName),
                 // segment to avoid reader allocations
                 payload = writer.ToArraySegment()
             };
@@ -311,7 +300,7 @@ namespace Mirror
         [EditorBrowsable(EditorBrowsableState.Never)]
         public virtual bool InvokeRPC(int rpcHash, NetworkReader reader)
         {
-            return InvokeHandlerDelegate(rpcHash, MirrorInvokeType.ClientRpc, reader);
+            return RemoteCallHelper.InvokeHandlerDelegate(rpcHash, MirrorInvokeType.ClientRpc, reader, this);
         }
         #endregion
 
@@ -331,7 +320,7 @@ namespace Mirror
                 netId = netId,
                 componentIndex = ComponentIndex,
                 // type+func so Inventory.RpcUse != Equipment.RpcUse
-                functionHash = GetMethodHash(invokeClass, eventName),
+                functionHash = RemoteCallHelper.GetMethodHash(invokeClass, eventName),
                 // segment to avoid reader allocations
                 payload = writer.ToArraySegment()
             };
@@ -348,168 +337,10 @@ namespace Mirror
         [EditorBrowsable(EditorBrowsableState.Never)]
         public virtual bool InvokeSyncEvent(int eventHash, NetworkReader reader)
         {
-            return InvokeHandlerDelegate(eventHash, MirrorInvokeType.SyncEvent, reader);
+            return RemoteCallHelper.InvokeHandlerDelegate(eventHash, MirrorInvokeType.SyncEvent, reader, this);
         }
         #endregion
 
-        #region Code Gen Path Helpers
-        /// <summary>
-        /// Delegate for Command functions.
-        /// </summary>
-        /// <param name="obj"></param>
-        /// <param name="reader"></param>
-        public delegate void CmdDelegate(NetworkBehaviour obj, NetworkReader reader, NetworkConnection senderConnection);
-
-        protected class Invoker
-        {
-            public Type invokeClass;
-            public MirrorInvokeType invokeType;
-            public CmdDelegate invokeFunction;
-            public bool cmdIgnoreAuthority;
-
-            public bool AreEqual(Type invokeClass, MirrorInvokeType invokeType, CmdDelegate function)
-            {
-                return (this.invokeClass == invokeClass &&
-                     this.invokeType == invokeType &&
-                     invokeFunction == function);
-            }
-        }
-
-        public struct CommandInfo
-        {
-            public bool ignoreAuthority;
-        }
-
-        static readonly Dictionary<int, Invoker> cmdHandlerDelegates = new Dictionary<int, Invoker>();
-
-        // helper function register a Command/Rpc/SyncEvent delegate
-        [EditorBrowsable(EditorBrowsableState.Never)]
-        protected static void RegisterDelegate(Type invokeClass, string cmdName, MirrorInvokeType invokerType, CmdDelegate func, bool cmdIgnoreAuthority = false)
-        {
-            // type+func so Inventory.RpcUse != Equipment.RpcUse
-            int cmdHash = GetMethodHash(invokeClass, cmdName);
-
-            if (CheckIfDeligateExists(invokeClass, invokerType, func, cmdHash))
-                return;
-
-            Invoker invoker = new Invoker
-            {
-                invokeType = invokerType,
-                invokeClass = invokeClass,
-                invokeFunction = func,
-                cmdIgnoreAuthority = cmdIgnoreAuthority,
-            };
-
-            cmdHandlerDelegates[cmdHash] = invoker;
-
-            if (logger.LogEnabled())
-            {
-                string ingoreAuthorityMessage = invokerType == MirrorInvokeType.Command ? $" IgnoreAuthority:{cmdIgnoreAuthority}" : "";
-                logger.Log($"RegisterDelegate hash: {cmdHash} invokerType: {invokerType} method: {func.GetMethodName()}{ingoreAuthorityMessage}");
-            }
-        }
-
-        static bool CheckIfDeligateExists(Type invokeClass, MirrorInvokeType invokerType, CmdDelegate func, int cmdHash)
-        {
-            if (cmdHandlerDelegates.ContainsKey(cmdHash))
-            {
-                // something already registered this hash
-                Invoker oldInvoker = cmdHandlerDelegates[cmdHash];
-                if (oldInvoker.AreEqual(invokeClass, invokerType, func))
-                {
-                    // it's all right,  it was the same function
-                    return true;
-                }
-
-                logger.LogError($"Function {oldInvoker.invokeClass}.{oldInvoker.invokeFunction.GetMethodName()} and {invokeClass}.{func.GetMethodName()} have the same hash.  Please rename one of them");
-            }
-
-            return false;
-        }
-
-        [EditorBrowsable(EditorBrowsableState.Never)]
-        public static void RegisterCommandDelegate(Type invokeClass, string cmdName, CmdDelegate func, bool ignoreAuthority)
-        {
-            RegisterDelegate(invokeClass, cmdName, MirrorInvokeType.Command, func, ignoreAuthority);
-        }
-
-        [EditorBrowsable(EditorBrowsableState.Never)]
-        public static void RegisterRpcDelegate(Type invokeClass, string rpcName, CmdDelegate func)
-        {
-            RegisterDelegate(invokeClass, rpcName, MirrorInvokeType.ClientRpc, func);
-        }
-
-        [EditorBrowsable(EditorBrowsableState.Never)]
-        public static void RegisterEventDelegate(Type invokeClass, string eventName, CmdDelegate func)
-        {
-            RegisterDelegate(invokeClass, eventName, MirrorInvokeType.SyncEvent, func);
-        }
-
-        // we need a way to clean up delegates after tests
-        [EditorBrowsable(EditorBrowsableState.Never)]
-        internal static void ClearDelegates()
-        {
-            cmdHandlerDelegates.Clear();
-        }
-
-        static bool GetInvokerForHash(int cmdHash, MirrorInvokeType invokeType, out Invoker invoker)
-        {
-            if (cmdHandlerDelegates.TryGetValue(cmdHash, out invoker) && invoker != null && invoker.invokeType == invokeType)
-            {
-                return true;
-            }
-
-            // debug message if not found, or null, or mismatched type
-            // (no need to throw an error, an attacker might just be trying to
-            //  call an cmd with an rpc's hash)
-            if (logger.LogEnabled()) logger.Log("GetInvokerForHash hash:" + cmdHash + " not found");
-
-            return false;
-        }
-
-        // InvokeCmd/Rpc/SyncEventDelegate can all use the same function here
-        internal bool InvokeHandlerDelegate(int cmdHash, MirrorInvokeType invokeType, NetworkReader reader, NetworkConnectionToClient senderConnection = null)
-        {
-            if (GetInvokerForHash(cmdHash, invokeType, out Invoker invoker) && invoker.invokeClass.IsInstanceOfType(this))
-            {
-                invoker.invokeFunction(this, reader, senderConnection);
-
-                return true;
-            }
-            return false;
-        }
-
-        internal CommandInfo GetCommandInfo(int cmdHash)
-        {
-            if (GetInvokerForHash(cmdHash, MirrorInvokeType.Command, out Invoker invoker) && invoker.invokeClass.IsInstanceOfType(this))
-            {
-                return new CommandInfo
-                {
-                    ignoreAuthority = invoker.cmdIgnoreAuthority
-                };
-            }
-            return default;
-        }
-
-        [Obsolete("Use NetworkBehaviour.GetDelegate instead.")]
-        public static CmdDelegate GetRpcHandler(int cmdHash) => GetDelegate(cmdHash);
-
-        /// <summary>
-        /// Gets the handler function for a given hash
-        /// Can be used by profilers and debuggers
-        /// </summary>
-        /// <param name="cmdHash">rpc function hash</param>
-        /// <returns>The function delegate that will handle the command</returns>
-        public static CmdDelegate GetDelegate(int cmdHash)
-        {
-            if (cmdHandlerDelegates.TryGetValue(cmdHash, out Invoker invoker))
-            {
-                return invoker.invokeFunction;
-            }
-            return null;
-        }
-
-        #endregion
 
         #region Helpers
