+++ Assets/Mirror/Runtime/RemoteCallHelper.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d2cdbcbd1e377d6408a91acbec31ba16
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
